feat(buffer): extend RTC ring buffer 48→288 readings (4h→24h offline coverage)
Widens rtcHead/rtcCount from uint8_t to uint16_t to support capacity > 255.
Bumps RingBuffer MAGIC 0xCB50A003→0xCB50A004 since rtcBuf shifts in LP-SRAM
when the count/head fields grow.

Memory cost: 288 × 28B = 8064 bytes (was 1344B). Still well under the
ESP32-C6's 16KB LP-SRAM budget.

Coverage: a 24-hour outage (storm, ISP cable cut, broker maintenance) now
preserves all 5-min-cadence telemetry. Beyond 24h the oldest readings are
silently overwritten (FIFO drop-oldest).

+++ firmware/sensor-tag-wifi/src/ring_buffer.h
@@ -17,10 +17,10 @@ bool peekOldest(Reading& out);
 void popOldest();
 
 /// Number of readings currently stored.
-uint8_t size();
+uint16_t size();
 
 /// Maximum capacity (compile-time).
-uint8_t capacity();
+uint16_t capacity();
 
 /// Reset buffer. Call from cold boot only — RTC memory is uninitialized then.
 void initIfColdBoot();
